+++ lqr/lqr_carver_pub.h
@@ -76,15 +76,15 @@ struct _LqrCarver
   LqrCarverList *attached_list; /* list of attached carvers */
 
   gfloat rigidity;              /* rigidity value (can straighten seams) */
-  gdouble *rigidity_map;        /* the rigidity function */
-  gdouble *rigidity_mask;	/* the rigidity mask */
+  gfloat *rigidity_map;        /* the rigidity function */
+  gfloat *rigidity_mask;	/* the rigidity mask */
   gint delta_x;                 /* max displacement of seams (currently is only meaningful if 0 or 1 */
 
   void *rgb;                    /* array of rgb points */
   gint *vs;                     /* array of visibility levels */
-  gdouble *en;                  /* array of energy levels */
-  gdouble *bias;                /* bias mask */
-  gdouble *m;                   /* array of auxiliary energy values */
+  gfloat *en;                  /* array of energy levels */
+  gfloat *bias;                /* bias mask */
+  gfloat *m;                   /* array of auxiliary energy values */
   gint *least;                  /* array of pointers */
   gint *_raw;                   /* array of array-coordinates, for seam computation */
   gint **raw;                   /* array of array-coordinates, for seam computation */
